Fix $0 fare: change fareEstimate from Decimal to Double for JSON compat
Root cause: Swift's JSONEncoder encodes Decimal via NSDecimalNumber
which can produce JSON that Android's json.getDouble() can't parse
(returns 0.0). Android uses Double for fare_estimate.

Fix: Changed RideOfferContent.fareEstimate from Decimal to Double.
The sats value is now passed as Double(fareSatsInt) which encodes
as a plain JSON number that Android reads correctly.

Also:
- Added sats display on fare card (≈ XXXX sats) for verification
- Added logging: "Sending offer: $XX USD = XXXX sats (BTC=$XXXXX)"

+++ RidestrSDK/Sources/RidestrSDK/Models/RideModels.swift
@@ -42,7 +42,8 @@ public enum RiderStage: String, Codable, Sendable {
 
 /// Content of a RoadFlare ride offer (Kind 3173, NIP-44 encrypted to driver).
 public struct RideOfferContent: Codable, Sendable {
-    public let fareEstimate: Decimal
+    /// Fare in satoshis (matching Android's Double encoding for JSON compatibility).
+    public let fareEstimate: Double
     public let destination: Location
     public let approxPickup: Location
     public let pickupRouteKm: Double?
@@ -67,7 +68,7 @@ public struct RideOfferContent: Codable, Sendable {
     }
 
     public init(
-        fareEstimate: Decimal,
+        fareEstimate: Double,
         destination: Location,
         approxPickup: Location,
         pickupRouteKm: Double? = nil,
